Create ODDPARITY.S
Designed and Developed an Odd Parity System in ARM ASSEMBLY that implements positive logic (input 0 when the key isn’t pressed and input 1 when the key is pressed). 

Odd Parity is an algorithm used to detect error during transmission in communication systems. 

Given three inputs as a 3-bit data value, such that if the input switch is pressed, that data bit is 1. This system has one output bit, creating a 4-bit value, such that the number of ones, considered as one 4-bit value will always be odd. The communication system sends the 4-bit value as a message (containing the 3-bit data plus parity), and the receiver could detect if one of the bits were to be flipped during transmission. 

Utilized Port Keys and LEDs in Assembly.